Major changes
For now I removed the animation script and archived it, I need to make a animation tree component and reference the animation system from that. But I have a good starting point now that I cleared off some code and can refer to the previous backup file I made. I need to do a few more things like add in more animations and work on a system that makes it better to export animations with the default rig, but this should be fine.. I hope